Date & datetime default setting to concern
Refactor date & datetime default setting functionality to DateDatetimeDefaults concern
